Lint: collapse cast to one line (ruff-format)

+++ vllm/models/kimi_k3/nvidia/mla.py
@@ -683,9 +683,7 @@ class MultiHeadLatentAttention(nn.Module, AttentionLayerBase):
                 # Backend dequantizes fp8 KV on load and takes a bf16 query
                 # (e.g. TRITON_MLA, the DSpark draft); undo the query
                 # quantization.
-                mqa_q = (mqa_q.to(torch.float32) * self._q_scale).to(
-                    ql_nope.dtype
-                )
+                mqa_q = (mqa_q.to(torch.float32) * self._q_scale).to(ql_nope.dtype)
             return mqa_q
         return fused_mla_decode_q_concat_kv_cache_insert(
             ql_nope,
